Auto-create a booking for new offenders
Every offender needs a booking record, so this commit auto-creates one
in the same way that the inbound movement is auto-created.

Also, change the automatic movement date to 4 days ago so that offenders
don't automatically go to the "Newly arrived" tab on MPC.

Also, auto-delete dependent offenders, bookings and movements. So when a
prison is deleted, everything else gets tidied up at the same time.